requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cloudinary==1.29.0
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
djongo==1.3.6
psycopg2==2.9.3
pymongo==4.2.0
six==1.16.0
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.26.11
